pbuilder: add crosscompile feature
This patch adds the feature of crossbuilding debian packages to the elbe
pbuilder. If 'elbe pbuilder --cross create' is called it will create a chroot
environment for crossbuilding. To use this crossbuild environment use
'elbe pbuilder build' with the --cross option.

Changelog:
 v2: applied Torben's suggestions
     add Error message for using '--cross' build without '--cross create'

 v3: applied Torben's suggestions
     create only the cross environment if 'pbuilder --cross' create is called.
     add error message for using '--cross build' without '--cross create'.
     add error message for using 'pbuilder build' if the environment was
     created with the cross option.

 v4: applied last changes suggested by Torben

Christian Teklenborg (3):
  pbuilder: adjust create_pbuilder and pdebuild_build
  pbuilder: add cross option to pbuilder
  doc: add cross parameter to pbuilder description

 docs/elbe-pbuilder.txt         |   7 +-
 elbepack/asyncworker.py        |  10 +--
 elbepack/commands/control.py   |   6 ++
 elbepack/commands/pbuilder.py  |   6 ++
 elbepack/daemons/soap/esoap.py |  12 ++--
 elbepack/db.py                 |  11 ++-
 elbepack/elbeproject.py        | 128 ++++++++++++++++++++++++---------
 elbepack/pbuilder.py           |  30 +++++++-
 elbepack/pbuilderaction.py     |  14 +++-
 elbepack/projectmanager.py     |  17 +++--
 elbepack/soapclient.py         |  13 ++--
 11 files changed, 189 insertions(+), 65 deletions(-)

--
2.20.1

------------------------------

Message: 3
Date: Thu,  7 May 2020 14:34:13 +0200
From: Christian Teklenborg <chris@linutronix.de>
To: elbe-devel@linutronix.de
Cc: Torben Hohn <torben.hohn@linutronix.de>
Subject: [elbe-devel] [PATCH v4 1/3] pbuilder: adjust create_pbuilder
	and	pdebuild_build
Message-ID: <20200507123415.5757-2-chris@linutronix.de>

Create a build environment for crossbuilding packages if the 'elbe pbuilder
--cross create' command is called. Now you can run 'elbe pbuilder --cross build'
to use the environment for crossbuilding. Change the create_pbuilder function
that it will create the correct chroot environment. Add the function
pbuilder_write_cross_config to write the correct config file for pbuilder.
Change the pbuilder_write_repo_hook function to take care of the different
directory names from the pbuilder environment.
Revise the pdebuild_build function to make sure that in case of --cross used
with the build command pbuilder runs with the '--host-arch' option. Throw an
error if the pbuilder environment was created with the --cross option but the
build command gets called without the --cross option and vice versa.

Signed-off-by: Christian Teklenborg <chris@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>