elbepack: elbeproject: Move host sysroot path generation.
The local repository was not added to the host syroot because of the path of
the host sysroot. The following condition was wrong.
elbepack/rfs.py: 122:         if os.path.exists(self.path + '/../repo/pool'):

So I move the host path builddir, then copy in the sdk path after building
process happened.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>