elbepack: elbeproject: Move host sysroot path generation.
The local repository was not added to the apt of the host syroot because of the
path of the host sysroot. The following condition was wrong.
elbepack/rfs.py: 122:         if os.path.exists(self.path + '/../repo/pool'):

The patch moves the host path creation into the builddir, then copies the
host chroot into the sdk path after the building process happened.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>