rpcaptcache.py: fix development package version in mark_install_devpkgs
The problem was the function mark_install_devpkgs installs all the dev
packages that match the same source name of an installed package but skip
if the version does not match.

This leads to miss development packages. If we build a X package with a
former version than the one available on the repo. If we add the package
with the version specified in the xml file this will not install the X-dev
package because the src_version is the recent version from the repo.

Update it to force the development package version and test if the candidate
is available.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>